DELETE DETIALS
DELETE DETAIL  LSIT TTILE

+++ lib/views/home/home.dart
@@ -84,9 +84,11 @@ class HomePageState extends State<HomePage> {
                       key: UniqueKey(),
                       background: Container(color: Colors.red),
                       onDismissed: (diretion) {
+                      if (_confirmDismiss()== true){
+                        DatabaseProvider.db
+                            .deleteZonaWithIddate(item.zona, item.date);
 
-                          DatabaseProvider.db
-                              .deleteZonaWithIddate(item.zona, item.date);
+                      }
 
 
                       },
